[LABS-483] Move wallet specific coin handlers out of WSM
