+++ chia/wallet/nft_wallet/nft_wallet.py
@@ -27,10 +27,15 @@ from chia.wallet.conditions import (
 from chia.wallet.derivation_record import DerivationRecord
 from chia.wallet.did_wallet import did_wallet_puzzles
 from chia.wallet.did_wallet.did_info import DIDInfo
+from chia.wallet.did_wallet.did_wallet import DIDWallet
 from chia.wallet.lineage_proof import LineageProof
 from chia.wallet.nft_wallet import nft_puzzle_utils
 from chia.wallet.nft_wallet.nft_info import NFTCoinInfo, NFTWalletInfo
-from chia.wallet.nft_wallet.nft_puzzle_utils import create_ownership_layer_puzzle, get_metadata_and_phs
+from chia.wallet.nft_wallet.nft_puzzle_utils import (
+    create_ownership_layer_puzzle,
+    get_metadata_and_phs,
+    get_new_owner_did,
+)
 from chia.wallet.nft_wallet.nft_puzzles import NFT_METADATA_UPDATER
 from chia.wallet.nft_wallet.uncurry_nft import NFTCoinData, UncurriedNFT
 from chia.wallet.outer_puzzles import AssetType, construct_puzzle, match_puzzle, solve_puzzle
@@ -43,7 +48,7 @@ from chia.wallet.trading.offer import OFFER_MOD, OFFER_MOD_HASH, NotarizedPaymen
 from chia.wallet.transaction_record import TransactionRecord
 from chia.wallet.uncurried_puzzle import uncurry_puzzle
 from chia.wallet.util.compute_additions import compute_additions
-from chia.wallet.util.wallet_types import WalletType
+from chia.wallet.util.wallet_types import WalletIdentifier, WalletType
 from chia.wallet.wallet import Wallet
 from chia.wallet.wallet_action_scope import WalletActionScope
 from chia.wallet.wallet_coin_record import WalletCoinRecord
@@ -51,6 +56,9 @@ from chia.wallet.wallet_info import WalletInfo
 from chia.wallet.wallet_nft_store import WalletNftStore
 from chia.wallet.wallet_protocol import GSTOptionalArgs, WalletProtocol
 from chia.wallet.wallet_spend_bundle import WalletSpendBundle
+
+if TYPE_CHECKING:
+    from chia.wallet.wallet_state_manager import WalletStateManager
 
 _T_NFTWallet = TypeVar("_T_NFTWallet", bound="NFTWallet")
 
@@ -262,6 +270,106 @@ class NFTWallet:
             minter_did,
             confirmed_height,
         )
+
+    @classmethod
+    async def identify(
+        cls,
+        wallet_state_manager: WalletStateManager,
+        nft_data: NFTCoinData,
+    ) -> WalletIdentifier | None:
+        """
+        Handle the new coin when it is a NFT
+        :param nft_data: all necessary data to process a NFT coin
+        :return: Wallet ID & Wallet Type
+        """
+        wallet_identifier = None
+        # DID ID determines which NFT wallet should process the NFT
+        new_did_id: bytes32 | None = None
+        old_did_id = None
+        # P2 puzzle hash determines if we should ignore the NFT
+        uncurried_nft: UncurriedNFT = nft_data.uncurried_nft
+        old_p2_puzhash = uncurried_nft.p2_puzzle.get_tree_hash()
+        _metadata, new_p2_puzhash = get_metadata_and_phs(
+            uncurried_nft,
+            nft_data.parent_coin_spend.solution,
+        )
+        if uncurried_nft.supports_did:
+            parsed_did_id = get_new_owner_did(
+                uncurried_nft, Program.from_serialized(nft_data.parent_coin_spend.solution)
+            )
+            old_did_id = uncurried_nft.owner_did
+            if parsed_did_id is None:
+                new_did_id = old_did_id
+            elif parsed_did_id == b"":
+                new_did_id = None
+            else:
+                new_did_id = parsed_did_id
+        wallet_state_manager.log.debug(
+            "Handling NFT: %s, old DID:%s, new DID:%s, old P2:%s, new P2:%s",
+            nft_data.parent_coin_spend,
+            old_did_id,
+            new_did_id,
+            old_p2_puzhash,
+            new_p2_puzhash,
+        )
+        new_derivation_record: (
+            DerivationRecord | None
+        ) = await wallet_state_manager.puzzle_store.get_derivation_record_for_puzzle_hash(new_p2_puzhash)
+        old_derivation_record: (
+            DerivationRecord | None
+        ) = await wallet_state_manager.puzzle_store.get_derivation_record_for_puzzle_hash(old_p2_puzhash)
+        if new_derivation_record is None and old_derivation_record is None:
+            wallet_state_manager.log.debug(
+                "Cannot find a P2 puzzle hash for NFT:%s, this NFT belongs to others.",
+                uncurried_nft.singleton_launcher_id.hex(),
+            )
+            return wallet_identifier
+        for nft_wallet in wallet_state_manager.wallets.copy().values():
+            if not isinstance(nft_wallet, NFTWallet):
+                continue
+            if nft_wallet.nft_wallet_info.did_id == old_did_id and old_derivation_record is not None:
+                wallet_state_manager.log.info(
+                    "Removing old NFT, NFT_ID:%s, DID_ID:%s",
+                    uncurried_nft.singleton_launcher_id.hex(),
+                    old_did_id,
+                )
+                if nft_data.parent_coin_state.spent_height is not None:
+                    await nft_wallet.remove_coin(
+                        nft_data.parent_coin_spend.coin, uint32(nft_data.parent_coin_state.spent_height)
+                    )
+                    is_empty = await nft_wallet.is_empty()
+                    has_did = False
+                    for did_wallet in wallet_state_manager.wallets.values():
+                        if not isinstance(did_wallet, DIDWallet):
+                            continue
+                        assert did_wallet.did_info.origin_coin is not None
+                        if did_wallet.did_info.origin_coin.name() == old_did_id:
+                            has_did = True
+                            break
+                    if is_empty and nft_wallet.did_id is not None and not has_did:
+                        wallet_state_manager.log.info(f"No NFT, deleting wallet {nft_wallet.did_id.hex()} ...")
+                        await wallet_state_manager.delete_wallet(nft_wallet.wallet_info.id)
+                        wallet_state_manager.wallets.pop(nft_wallet.wallet_info.id)
+            if nft_wallet.nft_wallet_info.did_id == new_did_id and new_derivation_record is not None:
+                wallet_state_manager.log.info(
+                    "Adding new NFT, NFT_ID:%s, DID_ID:%s",
+                    uncurried_nft.singleton_launcher_id.hex(),
+                    new_did_id,
+                )
+                wallet_identifier = WalletIdentifier.create(nft_wallet)
+
+        if wallet_identifier is None and new_derivation_record is not None:
+            # Cannot find an existed NFT wallet for the new NFT
+            wallet_state_manager.log.info(
+                "Cannot find a NFT wallet for NFT_ID: %s DID_ID: %s, creating a new one.",
+                uncurried_nft.singleton_launcher_id,
+                new_did_id,
+            )
+            new_nft_wallet: NFTWallet = await NFTWallet.create_new_nft_wallet(
+                wallet_state_manager, wallet_state_manager.main_wallet, did_id=new_did_id, name="NFT Wallet"
+            )
+            wallet_identifier = WalletIdentifier.create(new_nft_wallet)
+        return wallet_identifier
 
     async def add_coin(
         self,
